+++ R/NA_explicit_.R
@@ -1,13 +1,20 @@
-options( NA_explicit_ = ".NA")
+options( NA_explicit_ = "(NA)" ) 
 
 #' NA_explicit_
 #'
-#' Constant for categorical/character/factor variables
+#' Default replacement for missing values in categorical vectors.
 #' 
 #' @details 
 #' 
-#' `NA_explicit_` is an active binding to `getOptions('NA_explicit_`)` and 
-#' allows for a consistent, explicit `NA` values.
+#' `NA_explicit_` is used as a default replacement for categorical vectors. 
+#' It is an active binding to `getOptions('NA_explicit_')` and is exported
+#' to the callers namespace. 
+#' 
+#' To change the value of `NA_explicit` use:
+#' ```
+#' options( NA_explicit = new_value )
+#' ``` 
+#' You cannot directly set `NA_explicit_``
 #' 
 #' @seealso 
 #' 
@@ -17,8 +24,22 @@ options( NA_explicit_ = ".NA")
 #' @md
 #' @export
 
-NA_explicit_  <-
-  makeActiveBinding("NA_explicit_",
-                function(x) getOption("NA_explicit_"),
-                sys.frame(0)
+NA_explicit_ <- 
+  makeActiveBinding(
+      "NA_explicit_"
+    , function(x) {
+        val <- getOption("NA_explicit_")
+        # val <- structure(val, class = c("NA_explicit_", class(val)))
+        val
+      }
+    , sys.frame( sys.nframe() )
   )
+
+
+#' #' @rdname NA_explicit_
+#' #' @export 
+#' print.NA_explicit_ <- function(object) print( as.character( object ) ) 
+#' 
+#' #' @rdname NA_explicit_
+#' #' @export
+#' show.NA_explicit_ <- function(object) show( as.character( object ) )