+++ R/na_aggregate.R
@@ -1,47 +1,50 @@
-#' Replace NA by aggregation
+#' Replace NAs by aggregation
 #' 
-#' Generic function for replacing each \code{NA} with aggregated values. This
+#' Generic function for replacing each `NA` with aggregated values. This
 #' allows imputing by the overall mean, by monthly means, etc.
 #' 
 #' @param object an object.
-#' @param by a grouping variable corresponding to \code{object}, or a function
-#' to be applied to \code{time(object)} to generate the groups.
-#' @param \dots further arguments passed to \code{by} if \code{by} is a
+#' @param by a grouping variable corresponding to `object`, or a function
+#' to be applied to `time(object)` to generate the groups.
+#' @param \dots further arguments passed to `by` if `by` is a
 #' function.
 #' @param FUN function to apply to the non-missing values in each group defined
-#' by \code{by}.
-#' @param na_rm logical. Should any remaining \code{NA}s be removed?
-#' @param maxgap maximum number of consecutive \code{NA}s to fill. Any longer
+#' by `by`.
+#' @param na_rm logical. Should any remaining `NA`s be removed?
+#' @param maxgap maximum number of consecutive `NA`s to fill. Any longer
 #' gaps will be left unchanged.
-#' @return An object in which each \code{NA} in the input object is replaced by
-#' the mean (or other function) of its group, defined by \code{by}. This is
+#' @return An object in which each `NA` in the input object is replaced by
+#' the mean (or other function) of its group, defined by `by`. This is
 #' done for each series in a multi-column object. Common choices for the
 #' aggregation group are a year, a month, all calendar months, etc.
 #' 
 #' If a group has no non-missing values, the default aggregation function
-#' \code{mean} will return \code{NaN}. Specify \code{na_rm = TRUE} to omit such
+#' `mean` will return `NaN`. Specify \code{na_rm = TRUE} to omit such
 #' remaining missing values.
 #' 
 #' @seealso 
-#'   \code{\link{zoo}}
+#'   [na_mean()], [na_max()], [na_median()], etc. for similar
 #'   
 #' @examples
 #' 
-#' z <- zoo(c(1, NA, 3:9),
+#'   z <- zoo(c(1, NA, 3:9),
 #'          c(as.Date("2010-01-01") + 0:2,
 #'            as.Date("2010-02-01") + 0:2,
-#'            as.Date("2011-01-01") + 0:2))
-#' ## overall mean
-#' na_aggregate(z)
-#' ## group by months
-#' na_aggregate(z, as.yearmon)
-#' ## group by calendar months
-#' na_aggregate(z, months)
-#' ## group by years
-#' na_aggregate(z, format, "%Y")
+#'            as.Date("2011-01-01") + 0:2)
+#'          )
+#'          
+#'   ## overall mean
+#'   na_aggregate(z)
+#'   ## group by months
+#'   na_aggregate(z, as.yearmon)
+#'   ## group by calendar months
+#'   na_aggregate(z, months)
+#'   ## group by years
+#'   na_aggregate(z, format, "%Y")
 #' 
+#' @md
 #' @aliases na_aggregate na_aggregate.default
-#' @export na_aggregate
+#' @export 
 
 na_aggregate <- function(object, ...) UseMethod("na_aggregate")