+++ R/na_drop_cols.R
@@ -1,13 +1,18 @@
+#' na_drop_cols
+#' 
 #' Drop cols of a table whose values are all NA
 #' 
 #' @param data data-like object
 #' 
-#' Remove all cols whose only values are NA.
+#' @details
+#' `na_drop_cols` removes all cols whose only values are NA. It works on all 
+#' table-like objects.
 #' 
 #' @return 
-#'   An object of the same class as \code{data} with cols containing all 
-#'   \code{NA} values removed
+#'   An object of the same class as `data` with cols containing all 
+#'   `NA` values removed
 #' 
+#' @md
 #' @export
 
 na_drop_cols <- function(data) {