+++ R/na_drop_rows.R
@@ -1,17 +1,32 @@
+#' na_drop_rows
+#' 
 #' Drop rows of a table whose values are all NA
 #' 
-#' @param data data-like object
+#' @param tbl data-like object
 #' 
-#' Remove all rows whose only values are NA.
+#' @details 
+#' `na_drop_rows` removes all rows whose only values are NA. It works for all 
+#' table-like objects.
 #' 
 #' @return 
-#'   An object of the same class as \code{data} with rows containing all 
-#'   \code{NA} values removed
+#'   An object of the same class as `tbl` with rows containing all 
+#'   `NA` values removed
 #' 
+#' @examples 
+#' 
+#'   data(iris)
+#'   
+#'   tbl <- iris[1:5,]
+#'   tbl[1:3,] <- NA
+#'   tbl
+#'   
+#'   na_drop_rows(tbl) 
+#' 
+#' @md
 #' @export
 
-na_drop_rows <- function(data) {
-  data[ ! apply( data, 1, na_all ), ]
+na_drop_rows <- function(tbl) {
+  tbl[ ! apply( tbl, 1, na_all ), ]
 }
 
 #' @rdname na_drop_rows